+++ Tests/SentryTests/Integrations/Performance/IO/DataSentryTracingIntegrationTests.swift
@@ -20,7 +20,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
 
         init() {}
 
-        @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
         func getSut(testName: String, isSDKEnabled: Bool = true, isEnabled: Bool = true) throws -> Data {
             let fileManager = FileManager.default
             let tempDirUrl = URL(fileURLWithPath: NSTemporaryDirectory())
@@ -38,7 +37,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
                     // Configure options required by File I/O tracking integration
                     options.enableAutoPerformanceTracing = true
                     options.enableFileIOTracing = isEnabled
-                    options.setIntegrations(isEnabled ? [SentryFileIOTrackingIntegration.self] : [])
 
                     // Configure the tracing sample rate to record all traces
                     options.tracesSampleRate = 1.0
@@ -133,7 +131,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
 
     // MARK: - Data.init(contentsOfWithSentryTracing:)
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInitContentsOfWithSentryTracing_shouldTraceManually() throws {
         // -- Arrange --
         let expectedData = try fixture.getSut(testName: self.name)
@@ -163,7 +160,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertGreaterThan(endTimestamp.timeIntervalSince1970, startTimestamp.timeIntervalSince1970)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInitContentsOfWithSentryTracingWithOptions_shouldPassOptionsToSystemImplementation() throws {
         // -- Arrange --
         let expectedData = try fixture.getSut(testName: self.name)
@@ -191,7 +187,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(mappedSentryData, expectedData)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInitContentsOfWithSentryTracing_throwsError_shouldTraceManuallyWithErrorRethrow() throws {
         // -- Arrange --
         let _ = try fixture.getSut(testName: self.name)
@@ -218,7 +213,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertGreaterThan(endTimestamp.timeIntervalSince1970, startTimestamp.timeIntervalSince1970)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInitContentsOfWithSentryTracing_nonFileUrl_shouldNotTraceManually() throws {
         // -- Arrange --
         let _ = try fixture.getSut(testName: self.name)
@@ -232,7 +226,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(parentTransaction.children.count, 0)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInitContentsOfWithSentryTracing_trackerIsNotEnabled_shouldNotTraceManually() throws {
         // -- Arrange --
         let _ = try fixture.getSut(testName: self.name, isEnabled: false)
@@ -246,7 +239,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(parentTransaction.children.count, 0)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInitContentsOfWithSentryTracing_fileIsIgnored_shouldNotTraceManually() throws {
         // -- Arrange --
         let _ = try fixture.getSut(testName: self.name)
@@ -260,7 +252,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(parentTransaction.children.count, 0)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInitContentsOfWithSentryTracing_SDKIsNotStarted_shouldReadData() throws {
         // -- Arrange --
         let _ = try fixture.getSut(testName: self.name, isSDKEnabled: false)
@@ -273,7 +264,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(data, fixture.data)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInitContentsOfWithSentryTracing_SDKIsClosed_shouldReadData() throws {
         // -- Arrange --
         let _ = try fixture.getSut(testName: self.name)
@@ -289,7 +279,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
 
     // MARK: - Data.writeWithSentryTracing(to:)
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testWriteWithSentryTracing_shouldTraceManuallyWithErrorRethrow() throws {
         // -- Arrange --
         let sut: Data = try fixture.getSut(testName: self.name)
@@ -325,7 +314,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertGreaterThan(endTimestamp.timeIntervalSince1970, startTimestamp.timeIntervalSince1970)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testWriteWithSentryTracingWithOptions_shouldPassOptionsToSystemImplementation() throws {
         // -- Arrange --
         let sut: Data = try fixture.getSut(testName: self.name)
@@ -357,7 +345,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(writtenData, sut)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testWriteWithSentryTracing_throwsError_shouldTraceManuallyWithErrorRethrow() throws {
         // -- Arrange --
         let sut: Data = try fixture.getSut(testName: self.name)
@@ -384,7 +371,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertGreaterThan(endTimestamp.timeIntervalSince1970, startTimestamp.timeIntervalSince1970)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testWriteWithSentryTracing_nonFileUrl_shouldNotTraceManually() throws {
         // -- Arrange --
         let sut: Data = try fixture.getSut(testName: self.name)
@@ -397,7 +383,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(parentTransaction.children.count, 0)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testWriteWithSentryTracing_trackerIsNotEnabled_shouldNotTraceManually() throws {
         // -- Arrange --
         let sut: Data = try fixture.getSut(testName: self.name, isEnabled: false)
@@ -410,7 +395,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(parentTransaction.children.count, 0)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testWriteWithSentryTracing_fileIsIgnored_shouldNotTraceManually() throws {
         // -- Arrange --
         let sut: Data = try fixture.getSut(testName: self.name)
@@ -430,7 +414,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(parentTransaction.children.count, 0)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testWriteWithSentryTracing_SDKIsNotStarted_shouldWriteFile() throws {
         // -- Arrange --
         let sut: Data = try fixture.getSut(testName: self.name, isSDKEnabled: false)
@@ -449,7 +432,6 @@ class DataSentryTracingIntegrationTests: XCTestCase {
         XCTAssertEqual(writtenData, fixture.data)
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testWriteWithSentryTracing_SDKIsClosed_shouldWriteFile() throws {
         // -- Arrange --
         let sut: Data = try fixture.getSut(testName: self.name)