+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 ### Improvements
 
 - Replace deprecated SCNetworkReachability with NWPathMonitor (#6019)
+- Increase attachment max size to 100MB (#6537)
 
 ## 8.57.0